Reorganize for more nesting.
This removes the majority of structs from the crate root and puts all the controls in the controls module. It also adds a prelude, designed to be the only import from this library other than the specific controls used.
