--- ui/examples/rustlogo.rs
@@ -1,646 +1,0 @@
-//! An example app that displays the Rust logo.
-//!
-//! FIXME(pcwalton): This is a very broken Rust logo! It was halfheartedly autoconverted from SVG.
-
-extern crate ui;
-
-use std::rc::Rc;
-
-use ui::{UI, Area, AreaDrawParams, AreaHandler, Window};
-use ui::draw::{Brush, FillMode, Path, SolidBrush};
-
-struct RustLogoArea;
-
-trait PathExt {
-    fn arc2(
-        &self,
-        x_center: f64,
-        y_center: f64,
-        radius: f64,
-        start_angle: f64,
-        end_angle: f64,
-        anticlockwise: i32,
-    );
-}
-
-impl PathExt for Path {
-    fn arc2(
-        &self,
-        x_center: f64,
-        y_center: f64,
-        radius: f64,
-        start_angle: f64,
-        end_angle: f64,
-        anticlockwise: i32,
-    ) {
-        self.arc_to(
-            x_center,
-            y_center,
-            radius,
-            start_angle,
-            end_angle - start_angle,
-            anticlockwise != 0,
-        )
-    }
-}
-
-impl AreaHandler for RustLogoArea {
-    fn draw(&mut self, area: &Area, area_draw_params: &AreaDrawParams) {
-        let path = Path::new(FillMode::Winding);
-        path.new_figure(67.743, 31.035);
-        path.arc2(
-            0.0,
-            0.0,
-            3.108,
-            -3.1415926088863095,
-            -4.470348358154297e-8,
-            0,
-        );
-        path.arc2(
-            0.0,
-            0.0,
-            3.108,
-            4.4703483581542975e-8,
-            3.1415926088863095,
-            0,
-        );
-        path.new_figure(30.666, 59.175);
-        path.arc2(
-            0.0,
-            0.0,
-            3.108,
-            -3.1415926237874707,
-            -2.9802322387695312e-8,
-            0,
-        );
-        path.arc2(0.0, 0.0, 3.108, 0.0, 3.141592653589793, 0);
-        path.new_figure(104.819, 59.32);
-        path.arc2(
-            0.0,
-            0.0,
-            3.108,
-            -3.1415926088863095,
-            -4.470348358154297e-8,
-            0,
-        );
-        path.arc2(
-            0.0,
-            0.0,
-            3.108,
-            4.4703483581542975e-8,
-            3.1415926088863095,
-            0,
-        );
-        path.new_figure(39.663, 63.578);
-        path.bezier_to(
-            41.092999999999996,
-            62.943000000000005,
-            41.739,
-            61.267,
-            41.104,
-            59.834,
-        );
-        path.line_to(39.725, 56.716);
-        path.line_to(45.148, 56.716);
-        path.line_to(45.148, 81.16);
-        path.line_to(34.207, 81.16);
-        path.arc2(0.0, 0.0, 38.265, 2.86422996073528, 3.2497157630874733, 0);
-        path.close_figure();
-        path.new_figure(55.653000000000006, 67.154);
-        path.line_to(55.653000000000006, 59.949);
-        path.line_to(68.56700000000001, 59.949);
-        path.bezier_to(
-            69.23400000000001,
-            59.949,
-            73.277,
-            60.72,
-            73.277,
-            63.742999999999995,
-        );
-        path.bezier_to(
-            73.277,
-            66.253,
-            70.176,
-            67.15299999999999,
-            67.626,
-            67.15299999999999,
-        );
-        path.close_figure();
-        path.new_figure(44.717, 102.972);
-        path.arc2(0.0, 0.0, 3.108, 3.141592653589793, 6.283185307179586, 0);
-        path.arc2(0.0, 0.0, 3.108, 0.0, 3.141592653589793, 0);
-        path.new_figure(90.768, 103.11699999999999);
-        path.arc2(
-            0.0,
-            0.0,
-            3.108,
-            -3.1415926088863095,
-            -4.470348358154297e-8,
-            0,
-        );
-        path.arc2(
-            0.0,
-            0.0,
-            3.108,
-            4.4703483581542975e-8,
-            3.1415926088863095,
-            0,
-        );
-        path.new_figure(91.729, 96.06899999999999);
-        path.bezier_to(
-            90.198,
-            95.74099999999999,
-            88.692,
-            96.71499999999999,
-            88.364,
-            98.249,
-        );
-        path.line_to(86.804, 105.529);
-        path.arc2(0.0, 0.0, 38.265, 1.1454723568150609, 2.005709393175716, 0);
-        path.line_to(53.334, 98.09599999999999);
-        path.bezier_to(
-            53.006,
-            96.564,
-            51.5,
-            95.588,
-            49.970000000000006,
-            95.91699999999999,
-        );
-        path.line_to(43.543000000000006, 97.29699999999998);
-        path.arc2(0.0, 0.0, 38.265, 2.370885356807905, 2.5052253278412, 0);
-        path.line_to(71.492, 93.37999999999998);
-        path.bezier_to(
-            71.846,
-            93.37999999999998,
-            72.08200000000001,
-            93.31599999999999,
-            72.08200000000001,
-            92.99399999999999,
-        );
-        path.line_to(72.08200000000001, 81.932);
-        path.bezier_to(72.08200000000001, 81.61, 71.846, 81.546, 71.492, 81.546);
-        path.line_to(62.346000000000004, 81.546);
-        path.line_to(62.346000000000004, 74.534);
-        path.line_to(72.238, 74.534);
-        path.bezier_to(73.141, 74.534, 77.066, 74.792, 78.321, 79.80900000000001);
-        path.bezier_to(
-            78.714,
-            81.35200000000002,
-            79.577,
-            86.37100000000001,
-            80.167,
-            87.97800000000001,
-        );
-        path.bezier_to(
-            80.755,
-            89.78000000000002,
-            83.149,
-            93.38000000000001,
-            85.7,
-            93.38000000000001,
-        );
-        path.line_to(101.846, 93.38000000000001);
-        path.arc2(0.0, 0.0, 38.265, 0.6416547285438904, 0.7834412266197291, 0);
-        path.close_figure();
-        path.new_figure(115.667, 68.275);
-        path.arc2(
-            0.0,
-            0.0,
-            38.265,
-            -0.09911109621689877,
-            0.07472573234517861,
-            0,
-        );
-        path.line_to(111.822, 74.918);
-        path.bezier_to(111.429, 74.918, 111.271, 75.176, 111.271, 75.561);
-        path.line_to(111.271, 77.364);
-        path.bezier_to(111.271, 81.608, 108.878, 82.531, 106.781, 82.766);
-        path.bezier_to(
-            104.784,
-            82.991,
-            102.57000000000001,
-            81.93,
-            102.29700000000001,
-            80.708,
-        );
-        path.bezier_to(
-            101.11900000000001,
-            74.082,
-            99.156,
-            72.667,
-            96.05600000000001,
-            70.222,
-        );
-        path.bezier_to(99.903, 67.779, 103.906, 64.175, 103.906, 59.35099999999999);
-        path.bezier_to(
-            103.906,
-            54.141999999999996,
-            100.33500000000001,
-            50.86099999999999,
-            97.90100000000001,
-            49.251999999999995,
-        );
-        path.bezier_to(
-            94.486,
-            47.001,
-            90.70500000000001,
-            46.55,
-            89.68500000000002,
-            46.55,
-        );
-        path.line_to(42.509, 46.55);
-        path.arc2(
-            0.0,
-            0.0,
-            38.265,
-            -2.4116088516416183,
-            -1.7575991851921633,
-            0,
-        );
-        path.line_to(68.703, 39.489);
-        path.bezier_to(
-            69.785,
-            40.622,
-            71.577,
-            40.663999999999994,
-            72.709,
-            39.580999999999996,
-        );
-        path.line_to(78.06400000000001, 34.458999999999996);
-        path.arc2(
-            0.0,
-            0.0,
-            38.265,
-            -1.3856307110991473,
-            -0.5182172818862707,
-            0,
-        );
-        path.line_to(100.59400000000001, 61.396);
-        path.bezier_to(
-            99.96100000000001,
-            62.829,
-            100.60700000000001,
-            64.505,
-            102.036,
-            65.14,
-        );
-        path.close_figure();
-        path.new_figure(111.179, 65.274);
-        path.line_to(111.054, 63.994);
-        path.line_to(114.83, 60.472);
-        path.bezier_to(115.598, 59.756, 115.31099999999999, 58.315, 114.329, 57.949);
-        path.line_to(109.502, 56.144);
-        path.line_to(109.124, 54.897999999999996);
-        path.line_to(112.13499999999999, 50.715999999999994);
-        path.bezier_to(
-            112.749,
-            49.86599999999999,
-            112.18499999999999,
-            48.50899999999999,
-            111.151,
-            48.33899999999999,
-        );
-        path.line_to(106.06099999999999, 47.51099999999999);
-        path.line_to(105.449, 46.36799999999999);
-        path.line_to(107.588, 41.67299999999999);
-        path.bezier_to(
-            108.026,
-            40.716999999999985,
-            107.21199999999999,
-            39.493999999999986,
-            106.16,
-            39.533999999999985,
-        );
-        path.line_to(100.994, 39.713999999999984);
-        path.line_to(100.178, 38.72399999999998);
-        path.line_to(101.365, 33.69199999999998);
-        path.bezier_to(
-            101.60499999999999,
-            32.66999999999998,
-            100.568,
-            31.63199999999998,
-            99.54599999999999,
-            31.87199999999998,
-        );
-        path.line_to(94.51499999999999, 33.05799999999998);
-        path.line_to(93.52299999999998, 32.241999999999976);
-        path.line_to(93.70399999999998, 27.075999999999976);
-        path.bezier_to(
-            93.74399999999999,
-            26.029999999999976,
-            92.51999999999998,
-            25.212999999999976,
-            91.56599999999997,
-            25.646999999999977,
-        );
-        path.line_to(86.87199999999997, 27.786999999999978);
-        path.line_to(85.72899999999997, 27.173999999999978);
-        path.line_to(84.89899999999997, 22.082999999999977);
-        path.bezier_to(
-            84.73099999999997,
-            21.050999999999977,
-            83.37299999999998,
-            20.486999999999977,
-            82.52299999999997,
-            21.098999999999975,
-        );
-        path.line_to(78.33799999999997, 24.109999999999975);
-        path.line_to(77.09399999999997, 23.732999999999976);
-        path.line_to(75.28899999999996, 18.904999999999976);
-        path.bezier_to(
-            74.92299999999996,
-            17.920999999999978,
-            73.48099999999995,
-            17.637999999999977,
-            72.76699999999995,
-            18.401999999999976,
-        );
-        path.line_to(69.24499999999995, 22.180999999999976);
-        path.line_to(67.96499999999995, 22.055999999999976);
-        path.line_to(65.24499999999995, 17.660999999999976);
-        path.bezier_to(
-            64.69499999999995,
-            16.770999999999976,
-            63.221999999999944,
-            16.770999999999976,
-            62.67399999999995,
-            17.660999999999976,
-        );
-        path.line_to(59.95399999999995, 22.055999999999976);
-        path.line_to(58.67299999999995, 22.180999999999976);
-        path.line_to(55.14999999999995, 18.401999999999976);
-        path.bezier_to(
-            54.43599999999995,
-            17.637999999999977,
-            52.99399999999995,
-            17.920999999999975,
-            52.62799999999995,
-            18.904999999999976,
-        );
-        path.line_to(50.82299999999995, 23.732999999999976);
-        path.line_to(49.57799999999995, 24.109999999999975);
-        path.line_to(45.393999999999956, 21.098999999999975);
-        path.bezier_to(
-            44.543999999999954,
-            20.484999999999975,
-            43.18499999999995,
-            21.050999999999977,
-            43.01699999999995,
-            22.082999999999977,
-        );
-        path.line_to(42.186999999999955, 27.173999999999978);
-        path.line_to(41.043999999999954, 27.786999999999978);
-        path.line_to(36.34999999999995, 25.646999999999977);
-        path.bezier_to(
-            35.39599999999995,
-            25.210999999999977,
-            34.171999999999954,
-            26.029999999999976,
-            34.211999999999954,
-            27.075999999999976,
-        );
-        path.line_to(34.39199999999995, 32.241999999999976);
-        path.line_to(33.399999999999956, 33.05799999999998);
-        path.line_to(28.368999999999957, 31.87199999999998);
-        path.bezier_to(
-            27.34699999999996,
-            31.63399999999998,
-            26.30899999999996,
-            32.66999999999998,
-            26.548999999999957,
-            33.69199999999998,
-        );
-        path.line_to(27.733999999999956, 38.723999999999975);
-        path.line_to(26.919999999999956, 39.71399999999998);
-        path.line_to(21.753999999999955, 39.53399999999998);
-        path.bezier_to(
-            20.711999999999954,
-            39.50399999999998,
-            19.890999999999956,
-            40.71699999999998,
-            20.324999999999957,
-            41.67299999999998,
-        );
-        path.line_to(22.464999999999957, 46.36799999999998);
-        path.line_to(21.851999999999958, 47.51099999999998);
-        path.line_to(16.761999999999958, 48.338999999999984);
-        path.bezier_to(
-            15.727999999999957,
-            48.506999999999984,
-            15.167999999999958,
-            49.865999999999985,
-            15.777999999999958,
-            50.71599999999999,
-        );
-        path.line_to(18.78899999999996, 54.89799999999999);
-        path.line_to(18.41099999999996, 56.14399999999999);
-        path.line_to(13.58299999999996, 57.94899999999999);
-        path.bezier_to(
-            12.602999999999959,
-            58.31499999999999,
-            12.31599999999996,
-            59.75599999999999,
-            13.08199999999996,
-            60.471999999999994,
-        );
-        path.line_to(16.85899999999996, 63.99399999999999);
-        path.line_to(16.73399999999996, 65.27399999999999);
-        path.line_to(12.339999999999959, 67.99399999999999);
-        path.bezier_to(
-            11.449999999999958,
-            68.54399999999998,
-            11.449999999999958,
-            70.01699999999998,
-            12.339999999999959,
-            70.56499999999998,
-        );
-        path.line_to(16.73399999999996, 73.28499999999998);
-        path.line_to(16.85899999999996, 74.56499999999998);
-        path.line_to(13.081999999999958, 78.08799999999998);
-        path.bezier_to(
-            12.315999999999958,
-            78.80199999999998,
-            12.602999999999959,
-            80.24199999999998,
-            13.582999999999958,
-            80.60999999999999,
-        );
-        path.line_to(18.41099999999996, 82.41499999999999);
-        path.line_to(18.78899999999996, 83.66099999999999);
-        path.line_to(15.77799999999996, 87.844);
-        path.bezier_to(
-            15.16599999999996,
-            88.696,
-            15.72899999999996,
-            90.05399999999999,
-            16.76299999999996,
-            90.22,
-        );
-        path.line_to(21.85199999999996, 91.048);
-        path.line_to(22.99699999999996, 88.908);
-        path.line_to(27.689999999999962, 180.0);
-        path.line_to(27.253999999999962, 180.0);
-        path.line_to(29.434999999999963, 180.0);
-        path.line_to(31.573999999999963, 180.0);
-        path.line_to(36.737999999999964, 180.0);
-        path.line_to(37.55399999999997, 180.0);
-        path.line_to(42.58699999999997, 180.0);
-        path.line_to(42.346999999999966, 180.0);
-        path.line_to(43.14499999999997, 180.0);
-        path.line_to(44.96499999999997, 180.0);
-        path.line_to(49.99599999999997, 180.0);
-        path.line_to(50.987999999999964, 180.0);
-        path.line_to(56.154999999999966, 180.0);
-        path.line_to(56.11499999999997, 180.0);
-        path.line_to(57.298999999999964, 180.0);
-        path.line_to(59.43699999999996, 180.0);
-        path.line_to(64.13099999999996, 180.0);
-        path.line_to(65.27399999999996, 180.0);
-        path.line_to(70.36199999999995, 180.0);
-        path.line_to(70.52999999999996, 180.0);
-        path.line_to(72.05699999999996, 180.0);
-        path.line_to(74.43399999999995, 180.0);
-        path.line_to(78.61599999999996, 180.0);
-        path.line_to(79.86199999999995, 180.0);
-        path.line_to(81.66699999999996, 180.0);
-        path.bezier_to(
-            82.03299999999996,
-            180.0,
-            83.47499999999997,
-            180.0,
-            84.18899999999996,
-            180.0,
-        );
-        path.line_to(87.71199999999996, 180.0);
-        path.line_to(88.99299999999997, 180.0);
-        path.line_to(91.71299999999997, 180.0);
-        path.bezier_to(
-            92.26099999999997,
-            180.0,
-            93.73399999999997,
-            180.0,
-            94.28399999999996,
-            180.0,
-        );
-        path.line_to(97.00399999999996, 180.0);
-        path.line_to(98.28399999999996, 180.0);
-        path.line_to(101.80599999999997, 180.0);
-        path.bezier_to(
-            102.51999999999997,
-            180.0,
-            103.96199999999997,
-            180.0,
-            104.32799999999997,
-            180.0,
-        );
-        path.line_to(106.13299999999998, 180.0);
-        path.line_to(107.37899999999998, 180.0);
-        path.line_to(111.56199999999998, 180.0);
-        path.bezier_to(
-            112.41199999999998,
-            180.0,
-            113.76999999999998,
-            180.0,
-            113.93799999999999,
-            180.0,
-        );
-        path.line_to(114.76799999999999, 180.0);
-        path.line_to(115.91099999999999, 180.0);
-        path.line_to(120.60499999999999, 180.0);
-        path.bezier_to(
-            121.55899999999998,
-            180.0,
-            122.78099999999999,
-            180.0,
-            122.743,
-            180.0,
-        );
-        path.line_to(122.56299999999999, 180.0);
-        path.line_to(123.55399999999999, 180.0);
-        path.line_to(128.58499999999998, 180.0);
-        path.bezier_to(
-            129.60699999999997,
-            180.0,
-            130.64399999999998,
-            180.0,
-            130.40399999999997,
-            180.0,
-        );
-        path.line_to(129.21899999999997, 180.0);
-        path.line_to(130.03299999999996, 180.0);
-        path.line_to(135.19899999999996, 180.0);
-        path.bezier_to(
-            136.24099999999996,
-            180.0,
-            137.06499999999997,
-            180.0,
-            136.62699999999995,
-            180.0,
-        );
-        path.line_to(134.48799999999994, 180.0);
-        path.line_to(135.09999999999994, 180.0);
-        path.line_to(140.18999999999994, 180.0);
-        path.bezier_to(
-            141.22599999999994,
-            180.0,
-            141.78799999999995,
-            180.0,
-            141.17399999999995,
-            180.0,
-        );
-        path.line_to(138.16299999999995, 180.0);
-        path.line_to(138.54099999999994, 180.0);
-        path.line_to(143.36799999999994, 180.0);
-        path.bezier_to(
-            144.34999999999994,
-            180.0,
-            144.63699999999994,
-            180.0,
-            143.86899999999994,
-            180.0,
-        );
-        path.line_to(140.09299999999993, 180.0);
-        path.line_to(140.21799999999993, 180.0);
-        path.line_to(144.61199999999994, 180.0);
-        path.bezier_to(
-            145.50199999999992,
-            180.0,
-            145.50299999999993,
-            180.0,
-            144.61299999999994,
-            180.0,
-        );
-        path.end();
-
-        let brush = Brush::Solid(SolidBrush {
-            r: 0.0,
-            g: 0.0,
-            b: 0.0,
-            a: 255.0,
-        });
-        area_draw_params.context.fill(&path, &brush);
-        area.show();
-        // FIXME (leotindall): This doesn't actually draw anything?
-    }
-}
-
-fn run(ui: Rc<UI>) {
-    let window = Window::new("Rust logo", 640, 480, true);
-    {
-        let ui = ui.clone();
-        window.on_closing(Box::new(move |_| {
-            ui.quit();
-            false
-        }));
-    }
-
-    let area = Area::new(Box::new(RustLogoArea));
-    window.set_child(area.into());
-    window.show();
-    ui.main();
-}
-
-fn main() {
-    let ui = Rc::new(UI::init().unwrap());
-    run(ui);
-}
