+++ ui/src/controls/mod.rs
@@ -1,87 +1,21 @@
-//! Functions and types related to widgets.
+//! Contains all available controls and related functionality.
 
-use draw;
 use ffi_utils::{self, Text};
 use libc::{c_int, c_void};
 use std::ffi::CString;
 use std::mem;
 use std::ptr;
-use ui_sys::{self, uiArea, uiAreaDrawParams, uiAreaHandler, uiAreaKeyEvent, uiAreaMouseEvent};
-use ui_sys::{uiBox, uiButton, uiCheckbox, uiColorButton, uiCombobox, uiEditableCombobox,
-             uiControl, uiDateTimePicker};
-use ui_sys::{uiEntry, uiFontButton, uiGroup, uiLabel, uiMultilineEntry, uiProgressBar};
+use ui_sys::{self, uiBox, uiButton, uiCheckbox, uiCombobox, uiControl,
+             uiDateTimePicker, uiEditableCombobox};
+use ui_sys::{uiEntry, uiGroup, uiLabel, uiMultilineEntry, uiProgressBar};
 use ui_sys::{uiRadioButtons, uiSeparator, uiSlider, uiSpinbox, uiTab};
 
-pub use ui_sys::uiExtKey as ExtKey;
-
-// Defines a new control, creating a Rust wrapper, a `Deref` implementation, and a destructor.
-// An example of use:
-//
-//      define_control!{ 
-//          /// Some documentation
-//          #[attribute(whatever="something")]
-//          control(Slider, uiSlider, ui_slider); 
-//      }
-#[macro_export]
-macro_rules! define_control {
-    // Match first any attributes (incl. doc comments) and then the actual invocation
-    {$(#[$attr:meta])* control($rust_type:ident, $ui_type:ident, $ui_field:ident);} => {
-        // Include all attributes
-        $(#[$attr])*
-        pub struct $rust_type {
-            $ui_field: *mut $ui_type,
-        }
-
-        impl ::std::ops::Deref for $rust_type {
-            type Target = ::controls::Control;
-
-            #[inline]
-            fn deref(&self) -> &::controls::Control {
-                // FIXME(pcwalton): $10 says this is undefined behavior. How do I make it not so?
-                unsafe {
-                    mem::transmute::<&$rust_type, &::controls::Control>(self)
-                }
-            }
-        }
-
-        impl Drop for $rust_type {
-            #[inline]
-            fn drop(&mut self) {
-                // For now this does nothing, but in the future, when `libui` supports proper
-                // memory management, this will likely need to twiddle reference counts.
-            }
-        }
-
-        impl Clone for $rust_type {
-            #[inline]
-            fn clone(&self) -> $rust_type {
-                $rust_type {
-                    $ui_field: self.$ui_field,
-                }
-            }
-        }
-
-        impl Into<Control> for $rust_type {
-            #[inline]
-            fn into(self) -> Control {
-                unsafe {
-                    let control = Control::from_ui_control(self.$ui_field as *mut uiControl);
-                    mem::forget(self);
-                    control
-                }
-            }
-        }
-
-        impl $rust_type {
-            #[inline]
-            pub unsafe fn from_ui_control($ui_field: *mut $ui_type) -> $rust_type {
-                $rust_type {
-                    $ui_field: $ui_field
-                }
-            }
-        }
-    }
-}
+#[macro_use]
+mod create_macro;
+pub mod area;
+pub use self::area::Area;
+pub mod complex_selectors;
+pub use self::complex_selectors::{ColorButton, FontButton};
 
 /// A generic UI control. Any UI control can be dereferenced into this type.
 pub struct Control {
@@ -99,7 +33,9 @@ impl Drop for Control {
 impl Clone for Control {
     #[inline]
     fn clone(&self) -> Control {
-        Control { ui_control: self.ui_control }
+        Control {
+            ui_control: self.ui_control,
+        }
     }
 }
 
@@ -107,7 +43,9 @@ impl Control {
     /// Creates a new `Control` object from an existing `uiControl`.
     #[inline]
     pub unsafe fn from_ui_control(ui_control: *mut uiControl) -> Control {
-        Control { ui_control: ui_control }
+        Control {
+            ui_control: ui_control,
+        }
     }
 
     #[inline]
@@ -144,7 +82,7 @@ impl Control {
     }
 
     #[inline]
-    /// Set the parent control of this control, "moving" it to a new place in 
+    /// Set the parent control of this control, "moving" it to a new place in
     /// the UI tree or, if passed `None`, removing it.
     pub unsafe fn set_parent(&self, parent: Option<&Control>) {
         ffi_utils::ensure_initialized();
@@ -208,8 +146,7 @@ impl Control {
     }
 }
 
-
-define_control!{ 
+define_control!{
     /// A labeled clickable control which animates when clicked.
     control(Button, uiButton, ui_button);
 }
@@ -625,7 +562,7 @@ impl Spinbox {
     }
 }
 
-define_control!{ 
+define_control!{
     /// A control that displays a given value as a partial fill of a bar.
     control(ProgressBar, uiProgressBar, ui_progress_bar);
 }
@@ -643,7 +580,6 @@ impl ProgressBar {
         unsafe { ProgressBar::from_ui_control(ui_sys::uiNewProgressBar()) }
     }
 }
-
 
 define_control!{ /// A control that allows users to select a value by picking a location along a line.
     control(Slider, uiSlider, ui_slider);
@@ -890,381 +826,4 @@ impl MultilineEntry {
         ffi_utils::ensure_initialized();
         unsafe { MultilineEntry::from_ui_control(ui_sys::uiNewMultilineEntry()) }
     }
-}
-
-pub trait AreaHandler {
-    fn draw(&mut self, _area: &Area, _area_draw_params: &AreaDrawParams) {}
-    fn mouse_event(&mut self, _area: &Area, _area_mouse_event: &AreaMouseEvent) {}
-    fn mouse_crossed(&mut self, _area: &Area, _left: bool) {}
-    fn drag_broken(&mut self, _area: &Area) {}
-    fn key_event(&mut self, _area: &Area, _area_key_event: &AreaKeyEvent) -> bool {
-        true
-    }
-}
-
-#[repr(C)]
-struct RustAreaHandler {
-    ui_area_handler: uiAreaHandler,
-    trait_object: Box<AreaHandler>,
-}
-
-impl RustAreaHandler {
-    #[inline]
-    fn new(trait_object: Box<AreaHandler>) -> Box<RustAreaHandler> {
-        ffi_utils::ensure_initialized();
-        return Box::new(RustAreaHandler {
-            ui_area_handler: uiAreaHandler {
-                Draw: draw,
-                MouseEvent: mouse_event,
-                MouseCrossed: mouse_crossed,
-                DragBroken: drag_broken,
-                KeyEvent: key_event,
-            },
-            trait_object: trait_object,
-        });
-
-        extern "C" fn draw(
-            ui_area_handler: *mut uiAreaHandler,
-            ui_area: *mut uiArea,
-            ui_area_draw_params: *mut uiAreaDrawParams,
-        ) {
-            unsafe {
-                let area = Area::from_ui_area(ui_area);
-                let area_draw_params =
-                    AreaDrawParams::from_ui_area_draw_params(&*ui_area_draw_params);
-                (*(ui_area_handler as *mut RustAreaHandler))
-                    .trait_object
-                    .draw(&area, &area_draw_params);
-                mem::forget(area_draw_params);
-                mem::forget(area);
-            }
-        }
-
-        extern "C" fn mouse_event(
-            ui_area_handler: *mut uiAreaHandler,
-            ui_area: *mut uiArea,
-            ui_area_mouse_event: *mut uiAreaMouseEvent,
-        ) {
-            unsafe {
-                let area = Area::from_ui_area(ui_area);
-                let area_mouse_event =
-                    AreaMouseEvent::from_ui_area_mouse_event(&*ui_area_mouse_event);
-                (*(ui_area_handler as *mut RustAreaHandler))
-                    .trait_object
-                    .mouse_event(&area, &area_mouse_event);
-                mem::forget(area_mouse_event);
-                mem::forget(area);
-            }
-        }
-
-        extern "C" fn mouse_crossed(
-            ui_area_handler: *mut uiAreaHandler,
-            ui_area: *mut uiArea,
-            left: c_int,
-        ) {
-            unsafe {
-                let area = Area::from_ui_area(ui_area);
-                (*(ui_area_handler as *mut RustAreaHandler))
-                    .trait_object
-                    .mouse_crossed(&area, left != 0);
-                mem::forget(area);
-            }
-        }
-
-        extern "C" fn drag_broken(ui_area_handler: *mut uiAreaHandler, ui_area: *mut uiArea) {
-            unsafe {
-                let area = Area::from_ui_area(ui_area);
-                (*(ui_area_handler as *mut RustAreaHandler))
-                    .trait_object
-                    .drag_broken(&area);
-                mem::forget(area);
-            }
-        }
-
-        extern "C" fn key_event(
-            ui_area_handler: *mut uiAreaHandler,
-            ui_area: *mut uiArea,
-            ui_area_key_event: *mut uiAreaKeyEvent,
-        ) -> c_int {
-            unsafe {
-                let area = Area::from_ui_area(ui_area);
-                let area_key_event = AreaKeyEvent::from_ui_area_key_event(&*ui_area_key_event);
-                let result = (*(ui_area_handler as *mut RustAreaHandler))
-                    .trait_object
-                    .key_event(&area, &area_key_event);
-                mem::forget(area_key_event);
-                mem::forget(area);
-                result as c_int
-            }
-        }
-    }
-}
-
-/// A control which takes up space on which the application can draw custom content.
-define_control!{
-    /// A control which takes up space on which the application can draw custom content.
-    control(Area, uiArea, ui_area);
-}
-
-impl Area {
-    #[inline]
-    pub unsafe fn from_ui_area(ui_area: *mut uiArea) -> Area {
-        Area { ui_area: ui_area }
-    }
-
-    #[inline]
-    pub fn set_size(&self, width: i64, height: i64) {
-        ffi_utils::ensure_initialized();
-        unsafe { ui_sys::uiAreaSetSize(self.ui_area, width, height) }
-    }
-
-    #[inline]
-    pub fn queue_redraw_all(&self) {
-        ffi_utils::ensure_initialized();
-        unsafe { ui_sys::uiAreaQueueRedrawAll(self.ui_area) }
-    }
-
-    #[inline]
-    pub fn scroll_to(&self, x: f64, y: f64, width: f64, height: f64) {
-        ffi_utils::ensure_initialized();
-        unsafe { ui_sys::uiAreaScrollTo(self.ui_area, x, y, width, height) }
-    }
-
-    #[inline]
-    pub fn new(area_handler: Box<AreaHandler>) -> Area {
-        ffi_utils::ensure_initialized();
-        unsafe {
-            let mut rust_area_handler = RustAreaHandler::new(area_handler);
-            let area = Area::from_ui_control(ui_sys::uiNewArea(
-                &mut *rust_area_handler as *mut RustAreaHandler as
-                    *mut uiAreaHandler,
-            ));
-            mem::forget(rust_area_handler);
-            area
-        }
-    }
-
-    #[inline]
-    pub fn new_scrolling(area_handler: Box<AreaHandler>, width: i64, height: i64) -> Area {
-        ffi_utils::ensure_initialized();
-        unsafe {
-            let mut rust_area_handler = RustAreaHandler::new(area_handler);
-            let area = Area::from_ui_control(ui_sys::uiNewScrollingArea(
-                &mut *rust_area_handler as *mut RustAreaHandler as
-                    *mut uiAreaHandler,
-                width,
-                height,
-            ));
-            mem::forget(rust_area_handler);
-            area
-        }
-    }
-}
-
-pub struct AreaDrawParams {
-    pub context: draw::Context,
-
-    pub area_width: f64,
-    pub area_height: f64,
-
-    pub clip_x: f64,
-    pub clip_y: f64,
-    pub clip_width: f64,
-    pub clip_height: f64,
-}
-
-impl AreaDrawParams {
-    #[inline]
-    unsafe fn from_ui_area_draw_params(ui_area_draw_params: &uiAreaDrawParams) -> AreaDrawParams {
-        ffi_utils::ensure_initialized();
-        AreaDrawParams {
-            context: draw::Context::from_ui_draw_context(ui_area_draw_params.Context),
-            area_width: ui_area_draw_params.AreaWidth,
-            area_height: ui_area_draw_params.AreaHeight,
-            clip_x: ui_area_draw_params.ClipX,
-            clip_y: ui_area_draw_params.ClipY,
-            clip_width: ui_area_draw_params.ClipWidth,
-            clip_height: ui_area_draw_params.ClipHeight,
-        }
-    }
-}
-
-bitflags! {
-    pub flags Modifiers: u8 {
-        const MODIFIER_CTRL = 1 << 0,
-        const MODIFIER_ALT = 1 << 1,
-        const MODIFIER_SHIFT = 1 << 2,
-        const MODIFIER_SUPER = 1 << 3,
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-pub struct AreaMouseEvent {
-    pub x: f64,
-    pub y: f64,
-
-    pub area_width: f64,
-    pub area_height: f64,
-
-    pub down: i32,
-    pub up: i32,
-
-    pub count: i32,
-
-    pub modifiers: Modifiers,
-
-    pub held_1_to_64: u64,
-}
-
-impl AreaMouseEvent {
-    #[inline]
-    pub fn from_ui_area_mouse_event(ui_area_mouse_event: &uiAreaMouseEvent) -> AreaMouseEvent {
-        ffi_utils::ensure_initialized();
-        AreaMouseEvent {
-            x: ui_area_mouse_event.X,
-            y: ui_area_mouse_event.Y,
-            area_width: ui_area_mouse_event.AreaWidth,
-            area_height: ui_area_mouse_event.AreaHeight,
-            down: ui_area_mouse_event.Down,
-            up: ui_area_mouse_event.Up,
-            count: ui_area_mouse_event.Count,
-            modifiers: Modifiers::from_bits(ui_area_mouse_event.Modifiers as u8).unwrap(),
-            held_1_to_64: ui_area_mouse_event.Held1To64,
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug)]
-pub struct AreaKeyEvent {
-    pub key: u8,
-    pub ext_key: ExtKey,
-    pub modifier: Modifiers,
-    pub modifiers: Modifiers,
-    pub up: bool,
-}
-
-impl AreaKeyEvent {
-    #[inline]
-    pub fn from_ui_area_key_event(ui_area_key_event: &uiAreaKeyEvent) -> AreaKeyEvent {
-        ffi_utils::ensure_initialized();
-        AreaKeyEvent {
-            key: ui_area_key_event.Key as u8,
-            ext_key: ui_area_key_event.ExtKey,
-            modifier: Modifiers::from_bits(ui_area_key_event.Modifier as u8).unwrap(),
-            modifiers: Modifiers::from_bits(ui_area_key_event.Modifiers as u8).unwrap(),
-            up: ui_area_key_event.Up != 0,
-        }
-    }
-}
-
-define_control!{
-    /// A control which allows users to select a font.
-    control(FontButton, uiFontButton, ui_font_button);
-}
-
-impl FontButton {
-    /// Returns a new font.
-    #[inline]
-    pub fn font(&self) -> draw::text::Font {
-        ffi_utils::ensure_initialized();
-        unsafe {
-            draw::text::Font::from_ui_draw_text_font(ui_sys::uiFontButtonFont(self.ui_font_button))
-        }
-    }
-
-    #[inline]
-    pub fn on_changed(&self, callback: Box<FnMut(&FontButton)>) {
-        ffi_utils::ensure_initialized();
-        unsafe {
-            let mut data: Box<Box<FnMut(&FontButton)>> = Box::new(callback);
-            ui_sys::uiFontButtonOnChanged(
-                self.ui_font_button,
-                c_callback,
-                &mut *data as *mut Box<FnMut(&FontButton)> as *mut c_void,
-            );
-            mem::forget(data);
-        }
-
-        extern "C" fn c_callback(ui_font_button: *mut uiFontButton, data: *mut c_void) {
-            unsafe {
-                let font_button = FontButton::from_ui_control(ui_font_button);
-                mem::transmute::<*mut c_void, &mut Box<FnMut(&FontButton)>>(data)(&font_button);
-                mem::forget(font_button);
-            }
-        }
-    }
-
-    #[inline]
-    pub fn new() -> FontButton {
-        ffi_utils::ensure_initialized();
-        unsafe { FontButton::from_ui_control(ui_sys::uiNewFontButton()) }
-    }
-}
-
-define_control!{
-    /// A control which allows the user to select a color.
-    control(ColorButton, uiColorButton, ui_color_button);
-}
-
-impl ColorButton {
-    #[inline]
-    pub fn color(&self) -> Color {
-        ffi_utils::ensure_initialized();
-        unsafe {
-            let mut color: Color = mem::uninitialized();
-            ui_sys::uiColorButtonColor(
-                self.ui_color_button,
-                &mut color.r,
-                &mut color.g,
-                &mut color.b,
-                &mut color.a,
-            );
-            color
-        }
-    }
-
-    #[inline]
-    pub fn set_color(&self, color: &Color) {
-        ffi_utils::ensure_initialized();
-        unsafe {
-            ui_sys::uiColorButtonSetColor(self.ui_color_button, color.r, color.g, color.b, color.a)
-        }
-    }
-
-    #[inline]
-    pub fn on_changed(&self, callback: Box<FnMut(&ColorButton)>) {
-        ffi_utils::ensure_initialized();
-        unsafe {
-            let mut data: Box<Box<FnMut(&ColorButton)>> = Box::new(callback);
-            ui_sys::uiColorButtonOnChanged(
-                self.ui_color_button,
-                c_callback,
-                &mut *data as *mut Box<FnMut(&ColorButton)> as *mut c_void,
-            );
-            mem::forget(data);
-        }
-
-        extern "C" fn c_callback(ui_color_button: *mut uiColorButton, data: *mut c_void) {
-            unsafe {
-                let color_button = ColorButton::from_ui_control(ui_color_button);
-                mem::transmute::<*mut c_void, &mut Box<FnMut(&ColorButton)>>(data)(&color_button);
-                mem::forget(color_button)
-            }
-        }
-    }
-
-    #[inline]
-    pub fn new() -> ColorButton {
-        ffi_utils::ensure_initialized();
-        unsafe { ColorButton::from_ui_control(ui_sys::uiNewColorButton()) }
-    }
-}
-
-/// A RGBA color.
-#[derive(Copy, Clone, PartialEq, Debug)]
-pub struct Color {
-    r: f64,
-    g: f64,
-    b: f64,
-    a: f64,
 }
