+++ ui/src/lib.rs
@@ -5,21 +5,21 @@
 //!
 //! ```
 //! extern crate ui;
-//! use std::rc::Rc;
-//! use ui::{UI, Window, BoxControl, Button};
-//! 
+//! use ui::prelude::*;
+//! use ui::controls::{BoxControl, Button};
+//!
 //! fn main() {
 //!     // Start up the UI toolkit
 //!     let ui = Rc::new(UI::init().unwrap());
-//! 
+//!
 //!     // Create a new window, 200x100, titled "Test Window"
 //!     // and put it in an Rc so it can be passed into callback functions.
 //!     let main_window = Rc::new(Window::new("Test App", 200, 100, true));
-//! 
+//!
 //!     // Add margins around the edge of the window, making it look much nicer.
 //!     main_window.set_margined(true);
-//! 
-//!     // Adding this callback means that when this window closes, 
+//!
+//!     // Adding this callback means that when this window closes,
 //!     // the `ui::main` function returns.
 //!     // This should be added to the primary window of any application.
 //!     {
@@ -29,7 +29,7 @@
 //!             false
 //!         }));
 //!     }
-//! 
+//!
 //!     // Create a button that opens a dialog box.
 //!     let button = Button::new("Button");
 //!     {
@@ -42,7 +42,7 @@
 //!             main_window.msg_box("Button", "You clicked the button!");
 //!         }));
 //!     }
-//! 
+//!
 //!     // Create a button that quits the app.
 //!     let mut quit_button = Button::new("Quit");
 //!     {
@@ -57,14 +57,14 @@
 //!     vbox.append(quit_button.into(), false);
 //!     // Put the vertical layout into the window.
 //!     main_window.set_child(vbox.clone().into());
-//! 
+//!
 //!     // Set the main window (and all its widgets) to visible.
 //!     main_window.show();
-//! 
+//!
 //!     // Just for testing, quit the app before it even starts.
 //!     // Otherwise the test would never end!
 //!     ui.quit();
-//! 
+//!
 //!     // Run the app.
 //!     ui.main();
 //! }
@@ -75,22 +75,17 @@ extern crate bitflags;
 extern crate libc;
 extern crate ui_sys;
 
-pub use controls::{Area, AreaMouseEvent, AreaDrawParams, AreaKeyEvent, AreaHandler, BoxControl,
-                   Button, Checkbox, ColorButton};
-pub use controls::{Combobox, EditableCombobox, Control, DateTimePicker, Entry, FontButton, Group,
-                   Label};
-pub use controls::{MultilineEntry, ProgressBar, RadioButtons, Separator, Slider, Spinbox, Tab};
 pub use ffi_utils::Text;
-pub use menus::{Menu, MenuItem};
 pub use ui::{InitError, UI};
-pub use windows::Window;
+pub use window::Window;
 pub use image::Image;
 
 #[macro_use]
-mod controls;
+pub mod controls;
 pub mod draw;
 pub mod ffi_utils;
-mod menus;
+pub mod menus;
+pub mod prelude;
 mod ui;
-mod windows;
+pub mod window;
 mod image;
