+++ ui/src/prelude.rs
@@ -1,0 +1,4 @@
+//! Exports a set of useful features which are meant to be glob-imported.
+pub use std::rc::Rc;
+pub use UI;
+pub use window::Window;
